Changes in the calendar UI

+++ ASE_Calendar.Console/ConsoleOptions/HandleStateMaschine.cs
@@ -73,7 +73,7 @@ namespace ASE_Calendar.ConsoleUI.ConsoleOptions
                     calendar.CreateCalendarCurrentMonth();
 
                     
-                    Console.WriteLine("Previous month: left arrow | Next month: right arrow");
+                    Console.WriteLine("\nPrevious month: left arrow | Next month: right arrow");
 
                     if (currentUser.UserDataRegistered.RoleId == 2)
                     {
@@ -104,32 +104,32 @@ namespace ASE_Calendar.ConsoleUI.ConsoleOptions
                         Console.Clear();
                         currentTime = calendar.CreateCalendarNextMonth(currentTime);
                     }
-                    if (input.Key == ConsoleKey.F1)
+                    if (input.Key == ConsoleKey.F1 && currentUser.UserDataRegistered.RoleId is 2 or 0)
                     {
                         _appointmentManager = new AppointmentManager(currentUser, currentTime);
                         _appointmentManager.CreateAppointment();
                     }
-                    if (input.Key == ConsoleKey.F2)
+                    if (input.Key == ConsoleKey.F2 && currentUser.UserDataRegistered.RoleId is 2 or 0)
                     {
                         _appointmentManager = new AppointmentManager(currentUser, currentTime);
                         _appointmentManager.LoadAppointments();
                     }
-                    if (input.Key == ConsoleKey.F3)
+                    if (input.Key == ConsoleKey.F3 && currentUser.UserDataRegistered.RoleId == 0)
                     {
                         _appointmentManager = new AppointmentManager(currentUser, currentTime);
                         _appointmentManager.DeleteAnAppointment();
                     }
-                    if (input.Key == ConsoleKey.F4)
+                    if (input.Key == ConsoleKey.F4 && currentUser.UserDataRegistered.RoleId == 0)
                     {
                         _appointmentManager = new AppointmentManager(currentUser, currentTime);
                         _appointmentManager.LoadAllAppointments();
                     }
-                    if (input.Key == ConsoleKey.F5)
+                    if (input.Key == ConsoleKey.F5 && currentUser.UserDataRegistered.RoleId is 1 or 0)
                     {
                         _appointmentManager = new AppointmentManager(currentUser, currentTime);
                         _appointmentManager.ChangeDescriptionOfAnAppointment();
                     }
-                    if (input.Key == ConsoleKey.F6)
+                    if (input.Key == ConsoleKey.F6 && currentUser.UserDataRegistered.RoleId is 1 or 0)
                     {
                         _appointmentManager = new AppointmentManager(currentUser, currentTime);
                         _appointmentManager.ChangeDateOfAnAppointment();
